requirements.txt
python-dotenv
google-generativeai
colorama
PyYAML
flake8
pylint
openai